requirements.txt
tensorflow-cpu==2.4.*
tensorflow-gpu==2.4.*
tensorflow-addons==0.13.*
requests
keras
beautifulsoup4
pandas==1.0.0
numpy==1.19.5
lxml
loguru
grpcio
protobuf
